requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4 
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
langchain-openai
sentence-transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma
langsmith
streamlit
langchain-groq
fastapi
uvicorn
langserve